Refactor funcfile into its own package to allow use in langs
