+++ lambda.go
@@ -18,6 +18,8 @@ import (
 	"github.com/moby/moby/pkg/jsonmessage"
 	"github.com/urfave/cli"
 	yaml "gopkg.in/yaml.v2"
+
+	"github.com/fnproject/cli/funcfile"
 )
 
 var runtimes = map[string]string{
@@ -178,7 +180,7 @@ func createFunctionYaml(opts createImageOptions, functionName string) error {
 	strs := strings.Split(opts.Name, "/")
 	path := fmt.Sprintf("/%s", strs[1])
 
-	funcDesc := &Funcfile{
+	funcDesc := &funcfile.Funcfile{
 		Name:    opts.Name,
 		Version: "0.0.1",
 		Runtime: opts.Base,
